+++ src/main/java/com/sparta/invisible_project/service/CommentService.java
@@ -10,6 +10,7 @@ import com.sparta.invisible_project.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -30,7 +31,7 @@ public class CommentService {
     }
 
     // 댓글 상세 조회
-    public ResponseDto<?> findComment(long id) {
+    public ResponseDto<?> findComment(Long id) {
 
         Optional<Comment> comment = commentRepository.findById(id);
 
@@ -38,9 +39,8 @@ public class CommentService {
     }
 
     // 댓글 등록
-    public ResponseDto<?> createComment(CommentDto commentDto, long boardId) {
+    public ResponseDto<?> createComment(CommentDto commentDto, Long boardId, Member member) {
 
-        Member member = new Member();
         Board board = boardRepository.findById(boardId).orElse(null);
         Comment comment = new Comment(commentDto.getBody(), board, member);
 
@@ -50,20 +50,30 @@ public class CommentService {
     }
 
     // 댓글 수정
-    public ResponseDto<?> updateComment(CommentDto commentDto, long boardId, long commentId) {
+    @Transactional
+    public ResponseDto<?> updateComment(CommentDto commentDto, Long boardId, Long commentId, Member member) {
 
-        // 인증 필요
         Comment updateComment = commentRepository.findById(commentId).orElse(null);
+
+        if(updateComment.getMember().equals(member)){
+            return ResponseDto.fail(member.getUsername(), "댓글 작성자와 로그인한 사용자가 일치하지 않습니다. (UPDATE)");
+        }
+
         updateComment.update(commentDto.getBody());
 
         return ResponseDto.success(updateComment);
     }
 
     // 댓글 삭제
-    public ResponseDto<?> deleteComment(long boardId, long commentId) {
+    @Transactional
+    public ResponseDto<?> deleteComment(Long boardId, Long commentId, Member member) {
 
-        // 인증 필요
         Comment deleteComment = commentRepository.findById(commentId).orElse(null);
+
+        if(deleteComment.getMember().equals(member)){
+            return ResponseDto.fail(member.getUsername(), "댓글 작성자와 로그인한 사용자가 일치하지 않습니다. (DELETE)");
+        }
+
         commentRepository.deleteById(commentId);
 
         return ResponseDto.success(deleteComment);
